trainers: rename trainerERA5 → trainerERA5gen1
Consistent with train_gen1.py / train_gen2.py convention.
No external consumers — registry and tests are the only callers,
both updated here.
